+++ src/nom_parser/parser/parse_let.rs
@@ -1,42 +1,20 @@
-use crate::nom_parser::{
-    ast::{Expression, Statement},
-    token as Token,
-};
+use crate::nom_parser::parser::tools::parse_value;
+use crate::nom_parser::{ast::Statement, token as Token};
 
+use nom::sequence::preceded;
 use nom::{
-    branch::alt,
-    bytes::complete::{tag, take_until},
-    character::complete::{alphanumeric1, digit1, space0, space1},
-    sequence::{delimited, tuple},
+    bytes::complete::tag,
+    character::complete::{alphanumeric1, space0, space1},
+    sequence::tuple,
     IResult,
 };
 
-fn parse_string(input: &str) -> IResult<&str, Expression> {
-    let (input, x) = delimited(tag("\""), take_until("\""), tag("\""))(input)?;
-    Ok((input, Expression::String(String::from(x))))
-}
-
-fn parse_number(input: &str) -> IResult<&str, Expression> {
-    let (input, x) = digit1(input)?;
-    Ok((input, Expression::Number(x.parse().unwrap())))
-}
-
-fn parse_boolean(input: &str) -> IResult<&str, Expression> {
-    let (input, x) = alt((tag("true"), tag("false")))(input)?;
-    Ok((input, Expression::Boolean(x.parse().unwrap())))
-}
-
-fn parse_value(input: &str) -> IResult<&str, Expression> {
-    let (input, x) = alt((parse_number, parse_string, parse_boolean))(input)?;
-    Ok((input, x))
-}
-
 pub fn parse_let(input: &str) -> IResult<&str, Statement> {
-    let (input, (..)) = tuple((space0, tag(Token::LET)))(input)?;
+    let (input, ..) = preceded(space0, tag(Token::LET))(input)?;
 
-    let (input, (_, x)) = tuple((space1, alphanumeric1))(input)?;
+    let (input, x) = preceded(space1, alphanumeric1)(input)?;
 
-    let (input, (..)) = tuple((space0, tag(Token::ASSIGN), space0))(input)?;
+    let (input, ..) = tuple((space0, tag(Token::ASSIGN), space0))(input)?;
 
     let (input, y) = parse_value(input)?;
 
@@ -52,6 +30,7 @@ pub fn parse_let(input: &str) -> IResult<&str, Statement> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::nom_parser::ast::Expression;
 
     #[test]
     fn test1() {
@@ -111,7 +90,7 @@ mod tests {
 
     #[test]
     fn test5() {
-        match parse_let("let x = \"Hello") {
+        match parse_let("let x = Hello") {
             Ok(..) => {
                 panic!("need to fail")
             }
